+++ src/main/java/uk/ac/ebi/eva/accession/variant/VariantAccessioningConfiguration.java
@@ -28,7 +28,7 @@ import uk.ac.ebi.eva.accession.variant.persistence.VariantAccessioningDatabaseSe
 import uk.ac.ebi.eva.accession.variant.persistence.VariantAccessioningRepository;
 import uk.ac.ebi.ampt2d.commons.accession.autoconfigure.EnableSpringDataContiguousIdService;
 import uk.ac.ebi.ampt2d.commons.accession.generators.monotonic.MonotonicAccessionGenerator;
-import uk.ac.ebi.ampt2d.commons.accession.generators.monotonic.persistence.service.ContiguousIdBlockService;
+import uk.ac.ebi.ampt2d.commons.accession.persistence.monotonic.service.ContiguousIdBlockService;
 
 @Configuration
 @EnableSpringDataContiguousIdService
